[component,core] re-organize modules to better break up dependencies
